requirements.txt
mediapipe
python-dotenv 
numpy
opencv-python